backend/requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
